gestion upload premium

+++ site/wp-content/plugins/affieasy/classes/class-affiliation-table-admin.php
@@ -27,26 +27,28 @@ class AffiliationTableAdmin
                 array(),
                 time());
         });
-
-        aff_fs()->add_action('after_uninstall', 'aff_fs_uninstall_cleanup');
     }
 
-    public function initialize()
-    {
-        if (!$this->dbManager->table_exists(Constants::TABLE_WEBSHOP)) {
-            $this->dbManager->create_table_webshop();
-        }
-
-        if (!$this->dbManager->table_exists(Constants::TABLE_TABLE)) {
-            $this->dbManager->create_table_table();
-        }
-    }
-
-    public function rollback()
+    public static function initialize()
     {
         $staticDbManager = DbManager::get_instance();
-        $staticDbManager->drop_table(Constants::TABLE_WEBSHOP);
-        $staticDbManager->drop_table(Constants::TABLE_TABLE);
+        if (!$staticDbManager->table_exists(Constants::TABLE_WEBSHOP)) {
+            $staticDbManager->create_table_webshop();
+        }
+
+        if (!$staticDbManager->table_exists(Constants::TABLE_TABLE)) {
+            $staticDbManager->create_table_table();
+        }
+    }
+
+    public static function rollback() {
+        if (!is_dir(ABSPATH . 'wp-content/plugins/affieasy') || !is_dir(ABSPATH . 'wp-content/plugins/affieasy-premium')) {
+            $staticDbManager = DbManager::get_instance();
+            $staticDbManager->drop_table(Constants::TABLE_WEBSHOP);
+            $staticDbManager->drop_table(Constants::TABLE_TABLE);
+        }
+
+        aff_fs()->_uninstall_plugin_event();
     }
 
     public function add_menus_page_affiliation_table()
@@ -76,7 +78,7 @@ class AffiliationTableAdmin
             __('Webshops', 'affieasy'),
             'manage_options',
             'affieasy-webshop',
-            array($this, 'display_webshop_list')
+            array($this, 'display_webshop_views')
         );
     }
 
@@ -96,7 +98,7 @@ class AffiliationTableAdmin
         }
     }
 
-    public function display_webshop_list()
+    public function display_webshop_views()
     {
         if (is_admin() && current_user_can('manage_options')) {
             $action = isset($_GET['action']) ? $_GET['action'] : null;
